Merge branch 'virtio_net-ethtool-validation'
Nikolay Aleksandrov says:

====================
virtio_net: better ethtool setting validation

This small set is a follow-up for the recent patches that added ethtool
get/set settings. Patch 1 changes the speed validation routine to check
if the speed is between 0 and INT_MAX (or SPEED_UNKNOWN) and patch 2 adds
port validation to virtio_net and better validation comment.

This set is on top of Michael's patch which explains that speeds from 0
to INT_MAX are valid:
http://patchwork.ozlabs.org/patch/578911/
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 18ac5590e9e37a423c4d26a83b657fcf2b832d5a ]