rtc: s3c: Document in binding that only s3c6410 needs a src clk
The S3C binding doc says that the RTC and RTC source clocks are required
but the S3C driver supports different HW IP and only the s3c6410 needs a
source clock.

Fix the binding explaining that the source clock is only needed for the
s3c6410-rtc compatible controller.

Reported-by: Krzysztof Kozlowski <k.kozlowski@samsung.com>
Signed-off-by: Javier Martinez Canillas <javier@osg.samsung.com>
Acked-by: Alexandre Belloni <alexandre.belloni@free-electrons.com>
Reviewed-by: Krzysztof Kozlowski <k.kozlowski@samsung.com>
Signed-off-by: Alexandre Belloni <alexandre.belloni@free-electrons.com>

[ upstream commit: 00c9329ebe2d19a13afd86ba75218f868f2788ba ]