Merge branch 'rocker-worlds'
Jiri Pirko says:

====================
rocker: do world split

This patchset allows new rocker worlds to be easily added in future.
Two new worlds are now under development: P4 and eBPF.

The main part of the patchset is the OF-DPA carve-out. It resuts in OF-DPA
specific file. Clean cut.

Note this patchset is based on my original attempt in October 2015.
I had to rebase, included all suggestions and did lot of small changes.
Main change to go with all-port-one-world approach. Port world is set according
to what is setup in HW. Not possible to change worlds from driver.

v1->v2:
  patch 12/13:
  - split port_init into pre-init and init
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 2bc6b4f401730b153a0d5d0396335740e6af9ba8 ]