Merge branch 'ipvlan-misc'
Mahesh Bandewar says:

====================
IPvlan misc patches

This is a collection of unrelated patches for IPvlan driver.
a. crub_skb() changes are added to ensure that the packets hit the
NF_HOOKS in masters' ns in L3 mode.
b. u16 change is bug fix while
c. the third patch is to group tx/rx variables in single cacheline
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: ea5b2f4406d1b00324f0f572e505360695c8f8be ]