Merge branch 'qed-mf-updates'
Yuval Mintz says:

====================
qed: Management firmware updates

This series contains several changes to driver interaction with the
management fw.
The biggest [& most significant] change here is a change in the locking
scheme and re-definition of the 'critical section' when accessing shared
resources toward the goal of interacting with the management firmware.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 2aa45984c2180eecbbc9cc4e1321243308f41a01 ]