Merge branch 'flowi6_tos'
Daniel Borkmann says:

====================
flowi6_tos fixes

This set is a follow-up to address Jiri's recent feedback [1] on
the flowi6_tos issue, that it is not used for IPv6 route lookups.
The three patches fix all current users of flowi6_tos and remove
the define to avoid any future confusion on this. Tested the vxlan
and geneve ones with IPv6 routing rules. For details, please see
individual patches.

[ As fixes are currently applied against net-next tree, I've rebased
  it against that. ]

Thanks!

  [1] http://patchwork.ozlabs.org/patch/592055/
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 1e32ec3c8020051432f0b1c2314c8df6057f1b9b ]